+++ src/dp_trans_table.h
@@ -11,14 +11,14 @@ struct state_idx;
 struct dp_trans_table
 {
     uint32_t  ntrans;       /**< Number of transitions. */
-    double*   score;        /**< Transition score. */
+    float*    score;        /**< Transition score. */
     uint32_t* source_state; /**< Source state. */
     uint32_t* offset;       /**< Maps (target state, local trans) to score
                               and source_state indices. */
 };
 
 int dp_trans_table_change(struct dp_trans_table* trans_tbl, uint32_t src_state, uint32_t tgt_state,
-                          double lprob);
+                          float lprob);
 struct dp_trans_table* dp_trans_table_create(struct mstate const* const* mstates, uint32_t nstates,
                                              struct state_idx* state_idx);
 void                   dp_trans_table_destroy(struct dp_trans_table const* transition);
@@ -26,8 +26,8 @@ void                   dp_trans_table_dump(struct dp_trans_table const* trans_tb
 static inline uint32_t dp_trans_table_ntrans(struct dp_trans_table const* trans_tbl,
                                              uint32_t                     tgt_state);
 struct dp_trans_table* dp_trans_table_read(FILE* stream);
-static inline double   dp_trans_table_score(struct dp_trans_table const* trans_tbl,
-                                            uint32_t tgt_state, uint32_t trans);
+static inline float dp_trans_table_score(struct dp_trans_table const* trans_tbl, uint32_t tgt_state,
+                                         uint32_t trans);
 static inline uint32_t dp_trans_table_source_state(struct dp_trans_table const* trans_tbl,
                                                    uint32_t tgt_state, uint32_t trans);
 static inline uint32_t dp_trans_table_total_ntrans(struct dp_trans_table const* trans_tbl);
@@ -39,8 +39,8 @@ static inline uint32_t dp_trans_table_ntrans(struct dp_trans_table const* trans_
     return trans_tbl->offset[tgt_state + 1] - trans_tbl->offset[tgt_state];
 }
 
-static inline double dp_trans_table_score(struct dp_trans_table const* trans_tbl,
-                                          uint32_t tgt_state, uint32_t trans)
+static inline float dp_trans_table_score(struct dp_trans_table const* trans_tbl, uint32_t tgt_state,
+                                         uint32_t trans)
 {
     return trans_tbl->score[trans_tbl->offset[tgt_state] + trans];
 }
